refactoring of test frontend and fix library apth in ci workflow
